server/python-service/requirements.txt
cocoindex>=0.1.0
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
click>=8.0.0
